agnos/research_agent/requirements.txt
openai
duckduckgo-search
newspaper4k
lxml_html_clean
agno
python-dotenv
